+++ pkg/interpreter/interpreter_integration_test.go
@@ -286,6 +286,35 @@ var integrationTests = []struct {
 		wantCategory:   "",
 		config:         interpreter.DefaultConfig(),
 	},
+	// v0.27.0 regression tests
+	{
+		name:           "binary readwrite",
+		dir:            "binary_readwrite",
+		wantViolations: 0,
+		wantCategory:   "",
+		config:         interpreter.DefaultConfig(),
+	},
+	{
+		name:           "hash crc32",
+		dir:            "hash_crc32",
+		wantViolations: 0,
+		wantCategory:   "",
+		config:         interpreter.DefaultConfig(),
+	},
+	{
+		name:           "container list",
+		dir:            "container_list",
+		wantViolations: 0,
+		wantCategory:   "",
+		config:         interpreter.DefaultConfig(),
+	},
+	{
+		name:           "math big",
+		dir:            "math_big",
+		wantViolations: 0,
+		wantCategory:   "",
+		config:         interpreter.DefaultConfig(),
+	},
 	// v0.26.0 regression tests
 	{
 		name:           "time ticker",